Refactor message.rs into root.rs and messages.rs
Extract a new RootView, make MessagesView only concern about messages.
